add TestCase to common
